pyproject.toml
[project]
name = "flare-removal-2-0"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "gradio>=5.44.1",
    "numpy>=2.3.2",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "segmentation-models-pytorch>=0.5.0",
    "torch>=2.8.0",
    "uvicorn>=0.35.0",
]
